refactor: update home page
